data tests restructure
