+++ src/data/test/command_tests/create_user_tests.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-use crate::{commands::Command, data::data_tests_utils::*, session::Session};
+use crate::{commands::Command, data::test::data_tests_utils::*, session::Session};
 
 #[test]
 fn test_command_create_user_success() {
