+++ src/data/test/command_tests/get_user_tests.rs
@@ -1,6 +1,6 @@
 use std::str::FromStr;
 
-use crate::{commands::Command, data::data_tests_utils::*};
+use crate::{commands::Command, data::test::data_tests_utils::*};
 
 #[test]
 fn test_command_get_user() {
