+++ src/data/test/command_tests/revoke_tests.rs
@@ -1,17 +1,55 @@
 use std::str::FromStr;
 
-use crate::{commands::Command, data::data_tests_utils::*, session::Session};
+use crate::{commands::Command, data::test::data_tests_utils::*, session::Session};
 
 #[test]
-fn test_command_grant_number() {
+fn test_command_revoke_number() {
     let mut data = create_data_manager();
     let admin_session = create_session();
 
-    let cmd = Command::from_str("CREATE_USER user Password4 0").unwrap();
+    let cmd = Command::from_str("CREATE_USER user Password4 1").unwrap();
     let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
     assert_eq!(result, "OK".to_string());
 
-    let cmd = Command::from_str("GRANT user 1").unwrap();
+    let cmd = Command::from_str("REVOKE user 1").unwrap();
+    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
+    assert_eq!(result, "OK".to_string());
+
+    let cmd = Command::from_str("GET_USER user").unwrap();
+    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
+
+    assert_eq!(result, "User: user Permissions: 0".to_string());
+}
+
+#[test]
+fn test_command_revoke_name() {
+    let mut data = create_data_manager();
+    let admin_session = create_session();
+
+    let cmd = Command::from_str("CREATE_USER user Password4 8").unwrap();
+    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
+    assert_eq!(result, "OK".to_string());
+
+    let cmd = Command::from_str("REVOKE user USER_ADMIN").unwrap();
+    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
+    assert_eq!(result, "OK".to_string());
+
+    let cmd = Command::from_str("GET_USER user").unwrap();
+    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
+
+    assert_eq!(result, "User: user Permissions: 0".to_string());
+}
+
+#[test]
+fn test_command_revoke_name_multiple() {
+    let mut data = create_data_manager();
+    let admin_session = create_session();
+
+    let cmd = Command::from_str("CREATE_USER user Password4 11").unwrap();
+    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
+    assert_eq!(result, "OK".to_string());
+
+    let cmd = Command::from_str("REVOKE user USER_ADMIN GET").unwrap();
     let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
     assert_eq!(result, "OK".to_string());
 
@@ -22,45 +60,7 @@ fn test_command_grant_number() {
 }
 
 #[test]
-fn test_command_grant_name() {
-    let mut data = create_data_manager();
-    let admin_session = create_session();
-
-    let cmd = Command::from_str("CREATE_USER user Password4 0").unwrap();
-    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
-    assert_eq!(result, "OK".to_string());
-
-    let cmd = Command::from_str("GRANT user USER_ADMIN").unwrap();
-    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
-    assert_eq!(result, "OK".to_string());
-
-    let cmd = Command::from_str("GET_USER user").unwrap();
-    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
-
-    assert_eq!(result, "User: user Permissions: 8".to_string());
-}
-
-#[test]
-fn test_command_grant_name_multiple() {
-    let mut data = create_data_manager();
-    let admin_session = create_session();
-
-    let cmd = Command::from_str("CREATE_USER user Password4 0").unwrap();
-    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
-    assert_eq!(result, "OK".to_string());
-
-    let cmd = Command::from_str("GRANT user USER_ADMIN GET SET").unwrap();
-    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
-    assert_eq!(result, "OK".to_string());
-
-    let cmd = Command::from_str("GET_USER user").unwrap();
-    let (result, _) = data.handle_command(cmd, admin_session.clone()).unwrap();
-
-    assert_eq!(result, "User: user Permissions: 11".to_string());
-}
-
-#[test]
-fn test_command_grant_not_more_than_users() {
+fn test_command_revoke_not_more_than_users() {
     let mut data = create_data_manager();
     let admin_session = create_session();
 
@@ -77,7 +77,7 @@ fn test_command_grant_not_more_than_users() {
 
     assert_eq!(result, "OK".to_string());
 
-    let cmd = Command::from_str("GRANT user2 GET").unwrap();
+    let cmd = Command::from_str("REVOKE user2 GET").unwrap();
     let result = data.handle_command(cmd, user_session.clone()).unwrap_err();
 
     assert_eq!(result, "User does not have permission".to_string());
