+++ src/data/test/mod.rs
@@ -1,0 +1,4 @@
+mod auth_manager_tests;
+mod command_tests;
+mod data_tests_utils;
+mod key_tests;
